Add TinyGsmCommon
